Extract GuiTestBase and reuse in TestGui and TestGuiTotp
Moved members, init/cleanup and some help utility methods to the new class.
