+++ tests/gui/CMakeLists.txt
@@ -15,9 +15,9 @@
 
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}/..)
 
-add_unit_test(NAME testgui SOURCES TestGui.cpp ../util/TemporaryFile.cpp LIBS ${TEST_LIBRARIES})
+add_unit_test(NAME testgui SOURCES TestGui.cpp  GuiTestBase.cpp ../util/TemporaryFile.cpp LIBS ${TEST_LIBRARIES})
 add_unit_test(NAME testguipixmaps SOURCES TestGuiPixmaps.cpp LIBS ${TEST_LIBRARIES})
-add_unit_test(NAME testguitotp SOURCES TestGuiTotp.cpp ../util/TemporaryFile.cpp LIBS ${TEST_LIBRARIES})
+add_unit_test(NAME testguitotp SOURCES TestGuiTotp.cpp GuiTestBase.cpp ../util/TemporaryFile.cpp LIBS ${TEST_LIBRARIES})
 
 if(WITH_XC_BROWSER)
     add_unit_test(NAME testguibrowser SOURCES TestGuiBrowser.cpp ../util/TemporaryFile.cpp LIBS ${TEST_LIBRARIES})
