+++ tests/gui/GuiTestBase.cpp
@@ -1,0 +1,131 @@
+/*
+ *  Copyright (C) 2010 Felix Geyer <debfx@fobos.de>
+ *  Copyright (C) 2020 KeePassXC Team <team@keepassxc.org>
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation, either version 2 or (at your option)
+ *  version 3 of the License.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "GuiTestBase.h"
+#include "gui/Application.h"
+
+#include <QTest>
+
+#include "config-keepassx-tests.h"
+#include "crypto/Crypto.h"
+#include "gui/DatabaseTabWidget.h"
+#include "gui/FileDialog.h"
+#include "gui/MessageBox.h"
+#include "gui/PasswordWidget.h"
+
+void GuiTestBase::initTestCase()
+{
+    QVERIFY(Crypto::init());
+    Config::createTempFileInstance();
+    Application::bootstrap();
+
+    m_mainWindow.reset(new MainWindow());
+    m_tabWidget = m_mainWindow->findChild<DatabaseTabWidget*>("tabWidget");
+    m_statusBarLabel = m_mainWindow->findChild<QLabel*>("statusBarLabel");
+    m_mainWindow->show();
+    m_mainWindow->resize(1024, 768);
+}
+
+// Every test starts with resetting config settings and opening the temp database
+void GuiTestBase::init()
+{
+    // Reset config to defaults
+    config()->resetToDefaults();
+    // Disable autosave so we can test the modified file indicator
+    config()->set(Config::AutoSaveAfterEveryChange, false);
+    config()->set(Config::AutoSaveOnExit, false);
+    // Enable the tray icon so we can test hiding/restoring the windowQByteArray
+    config()->set(Config::GUI_ShowTrayIcon, true);
+    // Disable the update check first time alert
+    config()->set(Config::UpdateCheckMessageShown, true);
+    // Disable quick unlock
+    config()->set(Config::Security_QuickUnlock, false);
+    // Disable atomic saves to prevent transient errors on some platforms
+    config()->set(Config::UseAtomicSaves, false);
+    // Disable showing expired entries on unlock
+    config()->set(Config::GUI_ShowExpiredEntriesOnDatabaseUnlock, false);
+
+    // Copy the test database file to the temporary file
+    auto origFilePath = QDir(KEEPASSX_TEST_DATA_DIR).absoluteFilePath("NewDatabase.kdbx");
+    QVERIFY(m_dbFile.copyFromFile(origFilePath));
+
+    m_dbFileName = QFileInfo(m_dbFile.fileName()).fileName();
+    m_dbFilePath = m_dbFile.fileName();
+
+    // make sure window is activated or focus tests may fail
+    m_mainWindow->activateWindow();
+    QApplication::processEvents();
+
+    fileDialog()->setNextFileName(m_dbFilePath);
+    triggerAction("actionDatabaseOpen");
+
+    QApplication::processEvents();
+
+    m_dbWidget = m_tabWidget->currentDatabaseWidget();
+    auto* databaseOpenWidget = m_tabWidget->currentDatabaseWidget()->findChild<QWidget*>("databaseOpenWidget");
+    QVERIFY(databaseOpenWidget);
+    // editPassword is not QLineEdit anymore but PasswordWidget
+    auto* editPassword =
+        databaseOpenWidget->findChild<PasswordWidget*>("editPassword")->findChild<QLineEdit*>("passwordEdit");
+    QVERIFY(editPassword);
+    editPassword->setFocus();
+
+    QTest::keyClicks(editPassword, "a");
+    QTest::keyClick(editPassword, Qt::Key_Enter);
+
+    QTRY_VERIFY(!m_dbWidget->isLocked());
+    m_db = m_dbWidget->database();
+
+    QApplication::processEvents();
+}
+
+// Every test ends with closing the temp database without saving
+void GuiTestBase::cleanup()
+{
+    // DO NOT save the database
+    m_db->markAsClean();
+    MessageBox::setNextAnswer(MessageBox::No);
+    triggerAction("actionDatabaseClose");
+    QApplication::processEvents();
+    MessageBox::setNextAnswer(MessageBox::NoButton);
+
+    delete m_dbWidget;
+}
+
+void GuiTestBase::cleanupTestCase()
+{
+    m_dbFile.remove();
+}
+
+void GuiTestBase::triggerAction(const QString& name)
+{
+    auto* action = m_mainWindow->findChild<QAction*>(name);
+    QVERIFY(action);
+    QVERIFY(action->isEnabled());
+    action->trigger();
+    QApplication::processEvents();
+}
+
+void GuiTestBase::clickIndex(const QModelIndex& index,
+                             QAbstractItemView* view,
+                             Qt::MouseButton button,
+                             Qt::KeyboardModifiers stateKey)
+{
+    view->scrollTo(index);
+    QTest::mouseClick(view->viewport(), button, stateKey, view->visualRect(index).center());
+}
