+++ tests/gui/TestGui.h
@@ -19,6 +19,7 @@
 #ifndef KEEPASSX_TESTGUI_H
 #define KEEPASSX_TESTGUI_H
 
+#include "GuiTestBase.h"
 #include "gui/MainWindow.h"
 #include "util/TemporaryFile.h"
 
@@ -27,16 +28,11 @@ class DatabaseTabWidget;
 class DatabaseWidget;
 class QAbstractItemView;
 
-class TestGui : public QObject
+class TestGui : public GuiTestBase
 {
     Q_OBJECT
 
 private slots:
-    void initTestCase();
-    void init();
-    void cleanup();
-    void cleanupTestCase();
-
     void testSettingsDefaultTabOrder();
     void testCreateDatabase();
     void testMergeDatabase();
@@ -71,28 +67,15 @@ private:
     void addCannedEntries();
     void checkDatabase(const QString& filePath, const QString& expectedDbName);
     void checkDatabase(const QString& filePath = {});
-    void triggerAction(const QString& name);
     void dragAndDropGroup(const QModelIndex& sourceIndex,
                           const QModelIndex& targetIndex,
                           int row,
                           bool expectedResult,
                           const QString& expectedParentName,
                           int expectedPos);
-    void clickIndex(const QModelIndex& index,
-                    QAbstractItemView* view,
-                    Qt::MouseButton button,
-                    Qt::KeyboardModifiers stateKey = 0);
+
     void checkSaveDatabase();
     void checkStatusBarText(const QString& textFragment);
-
-    QScopedPointer<MainWindow> m_mainWindow;
-    QPointer<QLabel> m_statusBarLabel;
-    QPointer<DatabaseTabWidget> m_tabWidget;
-    QPointer<DatabaseWidget> m_dbWidget;
-    QSharedPointer<Database> m_db;
-    TemporaryFile m_dbFile;
-    QString m_dbFileName;
-    QString m_dbFilePath;
 };
 
 #endif // KEEPASSX_TESTGUI_H
