+++ tests/gui/TestGuiTotp.cpp
@@ -24,11 +24,6 @@
 #include <QTest>
 #include <QToolBar>
 
-#include "config-keepassx-tests.h"
-#include "crypto/Crypto.h"
-#include "gui/DatabaseTabWidget.h"
-#include "gui/FileDialog.h"
-#include "gui/MessageBox.h"
 #include "gui/PasswordWidget.h"
 #include "gui/TotpDialog.h"
 #include "gui/TotpSetupDialog.h"
@@ -51,90 +46,6 @@ int main(int argc, char* argv[])
     TestGuiTotp tc;
     QTEST_SET_MAIN_SOURCE_PATH
     return QTest::qExec(&tc, argc, argv);
-}
-
-void TestGuiTotp::initTestCase()
-{
-    QVERIFY(Crypto::init());
-    Config::createTempFileInstance();
-    Application::bootstrap();
-
-    m_mainWindow.reset(new MainWindow());
-    m_tabWidget = m_mainWindow->findChild<DatabaseTabWidget*>("tabWidget");
-    m_statusBarLabel = m_mainWindow->findChild<QLabel*>("statusBarLabel");
-    m_mainWindow->show();
-    m_mainWindow->resize(1024, 768);
-}
-
-// Every test starts with resetting config settings and opening the temp database
-void TestGuiTotp::init()
-{
-    // Reset config to defaults
-    config()->resetToDefaults();
-    // Disable autosave so we can test the modified file indicator
-    config()->set(Config::AutoSaveAfterEveryChange, false);
-    config()->set(Config::AutoSaveOnExit, false);
-    // Enable the tray icon so we can test hiding/restoring the windowQByteArray
-    config()->set(Config::GUI_ShowTrayIcon, true);
-    // Disable the update check first time alert
-    config()->set(Config::UpdateCheckMessageShown, true);
-    // Disable quick unlock
-    config()->set(Config::Security_QuickUnlock, false);
-    // Disable atomic saves to prevent transient errors on some platforms
-    config()->set(Config::UseAtomicSaves, false);
-    // Disable showing expired entries on unlock
-    config()->set(Config::GUI_ShowExpiredEntriesOnDatabaseUnlock, false);
-
-    // Copy the test database file to the temporary file
-    auto origFilePath = QDir(KEEPASSX_TEST_DATA_DIR).absoluteFilePath("NewDatabase.kdbx");
-    QVERIFY(m_dbFile.copyFromFile(origFilePath));
-
-    m_dbFileName = QFileInfo(m_dbFile.fileName()).fileName();
-    m_dbFilePath = m_dbFile.fileName();
-
-    // make sure window is activated or focus tests may fail
-    m_mainWindow->activateWindow();
-    QApplication::processEvents();
-
-    fileDialog()->setNextFileName(m_dbFilePath);
-    triggerAction("actionDatabaseOpen");
-
-    QApplication::processEvents();
-
-    m_dbWidget = m_tabWidget->currentDatabaseWidget();
-    auto* databaseOpenWidget = m_tabWidget->currentDatabaseWidget()->findChild<QWidget*>("databaseOpenWidget");
-    QVERIFY(databaseOpenWidget);
-    // editPassword is not QLineEdit anymore but PasswordWidget
-    auto* editPassword =
-        databaseOpenWidget->findChild<PasswordWidget*>("editPassword")->findChild<QLineEdit*>("passwordEdit");
-    QVERIFY(editPassword);
-    editPassword->setFocus();
-
-    QTest::keyClicks(editPassword, "a");
-    QTest::keyClick(editPassword, Qt::Key_Enter);
-
-    QTRY_VERIFY(!m_dbWidget->isLocked());
-    m_db = m_dbWidget->database();
-
-    QApplication::processEvents();
-}
-
-// Every test ends with closing the temp database without saving
-void TestGuiTotp::cleanup()
-{
-    // DO NOT save the database
-    m_db->markAsClean();
-    MessageBox::setNextAnswer(MessageBox::No);
-    triggerAction("actionDatabaseClose");
-    QApplication::processEvents();
-    MessageBox::setNextAnswer(MessageBox::NoButton);
-
-    delete m_dbWidget;
-}
-
-void TestGuiTotp::cleanupTestCase()
-{
-    m_dbFile.remove();
 }
 
 void TestGuiTotp::testTotp()
@@ -206,22 +117,4 @@ void TestGuiTotp::testTotp()
     qrCodeDialog->setFixedSize(800, 600);
     QVERIFY2(qrCodeWidget->geometry().width() == qrCodeWidget->geometry().height(), "Resized QR code is not square");
     QTest::keyClick(qrCodeDialog, Qt::Key_Escape);
-}
-
-void TestGuiTotp::triggerAction(const QString& name)
-{
-    auto* action = m_mainWindow->findChild<QAction*>(name);
-    QVERIFY(action);
-    QVERIFY(action->isEnabled());
-    action->trigger();
-    QApplication::processEvents();
-}
-
-void TestGuiTotp::clickIndex(const QModelIndex& index,
-                             QAbstractItemView* view,
-                             Qt::MouseButton button,
-                             Qt::KeyboardModifiers stateKey)
-{
-    view->scrollTo(index);
-    QTest::mouseClick(view->viewport(), button, stateKey, view->visualRect(index).center());
 }
